+++ test/sliver_tree/bulk_reentry_continuation_test.dart
@@ -15,7 +15,6 @@
 ///     members pop to full extent instantly.
 library;
 
-import 'package:flutter/animation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:widgets_extended/widgets_extended.dart';